// src/type/zhihu/column.d.ts
export type Record = {
  updated: 1543830773 | number
  description: '科学，以玩为本。' | string
  topics: [
    {
      url: 'https://www.zhihu.com/api/v4/topics/20010094' | string
      type: 'topic' | string
      id: '20010094' | string
      name: '黑科技' | string
    },
    {
      url: 'https://www.zhihu.com/api/v4/topics/19555490' | string
      type: 'topic' | string
      id: '19555490' | string
      name: '手机游戏' | string
    },
  ]
  can_manage: false | boolean
  intro: '公众号：狂丸科学' | string
  is_following: false | boolean
  url_token: 'kuangwan' | string
  id: 'kuangwan' | string
  articles_count: 371 | number
  accept_submission: false | boolean
  title: '狂丸科学' | string
  url: 'https://zhuanlan.zhihu.com/kuangwan' | string
  comment_permission: 'all' | string
  created: 1497252346 | number
  image_url: 'https://pic2.zhimg.com/v2-59c79c22006310ae562e83988491fcbb_b.jpg' | string
  author: {
    is_followed: false | boolean
    avatar_url_template: 'https://pic4.zhimg.com/v2-cba887679509294b7b929970e9038056_{size}.jpg' | string
    uid: '856466662591987712' | string
    user_type: 'people' | string
    is_following: false | boolean
    url_token: 'kuang-wan-35-41' | string
    id: '5f33ef832172f703f4434ca4e2b7a0b4' | string
    description: '微信公众号：狂丸科学，获取科学学习资源' | string
    name: '狂丸科学' | string
    is_advertiser: false | boolean
    headline: '微信公众号：狂丸科学，获取科学学习资源' | string
    gender: 1 | number
    url: '/people/5f33ef832172f703f4434ca4e2b7a0b4' | string
    avatar_url: 'https://pic4.zhimg.com/v2-cba887679509294b7b929970e9038056_l.jpg' | string
    is_org: false | boolean
    type: 'people' | string
  }
  followers: 18724 | number
  type: 'column' | string
}